requirements.txt
Flask==2.3.3
anthropic==0.25.1
python-dotenv==1.0.0
requests==2.32.5
